Fix game price readability
Fix steam game's price readability on steam store's landing page. Some games that is highlighted on top of the page, that have grey or light green or white cover art hinders the game's price visual. By adding a shadow under both the original and the final price of the games, hopefully this can improve the readability of the game's price information.